backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
motor==3.3.2
pydantic==2.5.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
faker==20.1.0
python-multipart==0.0.6
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
